[Radix Cache] Move TreeCore inspector into tests
